fix: 유저 활동 내역을 조회하기 위한 쿼리 오류 수정

+++ src/main/java/com/codeit/monew/comment/dto/CommentLikeActivityDto.java
@@ -12,7 +12,7 @@ public record CommentLikeActivityDto(
 	UUID commentUserId,
 	String commentUserNickname,
 	String commentContent,
-	int commentLikeCount,
+	long commentLikeCount,
 	Instant commentCreatedAt
 ) {
 }
